requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot>=11.1.0
psycopg2-binary
pyowm
feedparser
bs4
lxml
demjson
PyLyrics
wget
gTTS
telegram
zalgo_text
hurry.filesize
